+++ ksql-engine/src/main/java/io/confluent/ksql/KsqlContext.java
@@ -75,7 +75,8 @@ public class KsqlContext {
     final MutableFunctionRegistry functionRegistry = new InternalFunctionRegistry();
     UdfLoader.newInstance(ksqlConfig, functionRegistry, ".").load();
     final String serviceId = ksqlConfig.getString(KsqlConfig.KSQL_SERVICE_ID_CONFIG);
-    final String customMetricsTags = ksqlConfig.getString(KsqlConfig.KSQL_CUSTOM_METRICS_TAGS);
+    final Map<String, String> customMetricsTags =
+        ksqlConfig.getStringAsMap(KsqlConfig.KSQL_CUSTOM_METRICS_TAGS);
     final KsqlEngine engine = new KsqlEngine(
         serviceContext,
         processingLogContext,