+++ ksql-engine/src/test/java/io/confluent/ksql/integration/JsonFormatTest.java
@@ -98,7 +98,7 @@ public class JsonFormatTest {
         ProcessingLogContext.create(),
         new InternalFunctionRegistry(),
         ksqlConfig.getString(KsqlConfig.KSQL_SERVICE_ID_CONFIG),
-        ksqlConfig.getString(KsqlConfig.KSQL_CUSTOM_METRICS_TAGS));
+        ksqlConfig.getStringAsMap(KsqlConfig.KSQL_CUSTOM_METRICS_TAGS));
 
     topicClient = serviceContext.getTopicClient();
     metaStore = ksqlEngine.getMetaStore();